requirements.txt
langchain==0.3.26
flask==3.1.1 
# sentence-transformers==4.1.0
pypdf==5.6.1 
python-dotenv==1.1.0
langchain-google-genai==2.1.8
langchain-community==0.3.26
langchain-huggingface==0.3.1
pinecone-client==3.1.0
langchain-pinecone==0.2.8
gunicorn==21.2.0
-e .